+++ platform/fetch-native/src/index.ts
@@ -1,0 +1,5 @@
+import { setFetchImpl } from '@affinidi/platform-fetch'
+
+export const useNativeFetch = () => {
+  setFetchImpl(fetch)
+}