+++ sdk/core/test/integration/index.test.ts
@@ -1,6 +1,9 @@
 import 'mocha'
 import './env'
 
+import { useNodeFetch } from '@affinidi/platform-fetch-node'
+useNodeFetch()
+
 require('./migration')
 require('./services')
 require('./AffinityWallet.test')