+++ sdk/node/test/integration/index.test.ts
@@ -1,4 +1,9 @@
 import 'mocha'
+
+import { useNodeFetch } from '@affinidi/platform-fetch-node'
+
+useNodeFetch()
+
 import './env'
 
 require('./vc-signatures.test')